vmm: refactor vcpu.rs file
Break up vcpu into 3 files: one that has
common functionality for both architectures (mainly the
one in charge of signal handling) and 2 files per architecture.
